# Use zip_longest for more readable and faster code in the paired end iterator