--- a/src/dnaio/pairedend.py
+++ b/src/dnaio/pairedend.py
@@ -1,3 +1,4 @@
+import itertools
 from contextlib import ExitStack
 from os import PathLike
 from typing import Union, BinaryIO, Optional, Iterator, Tuple
@@ -47,26 +48,25 @@ class TwoFilePairedEndReader(PairedEndReader):
         """
         Iterate over the paired reads. Each item is a pair of Sequence objects.
         """
-        # Avoid usage of zip() below since it will consume one item too many.
+        # Avoid usage of zip() below since it will consume one item too many,
+        # when one of the iterators is exhausted. zip in python 3.10 has a
+        # 'strict' keyword that can be used to prevent this and throw an error,
+        # but it will take a long time for 3.10 or higher to be available on
+        # everyone's machine.
+        # Instead use zip_longest from itertools. This yields None if one of
+        # the iterators is exhausted. Checking for truthy values is speedy.
+        # So we can quickly check if the iterator is still yielding.
+        # This is faster than implementing a while loop with next calls,
+        # which requires expensive function lookups.
         it1, it2 = iter(self.reader1), iter(self.reader2)
-        while True:
-            try:
-                r1 = next(it1)
-            except StopIteration:
-                # End of file 1. Make sure that file 2 is also at end.
-                try:
-                    next(it2)
-                    raise FileFormatError(
-                        "Reads are improperly paired. There are more reads in "
-                        "file 2 than in file 1.",
-                        line=None,
-                    ) from None
-                except StopIteration:
-                    pass
-                break
-            try:
-                r2 = next(it2)
-            except StopIteration:
+        for r1, r2 in itertools.zip_longest(it1, it2):
+            if not r1:
+                raise FileFormatError(
+                    "Reads are improperly paired. There are more reads in "
+                    "file 2 than in file 1.",
+                    line=None,
+                ) from None
+            if not r2:
                 raise FileFormatError(
                     "Reads are improperly paired. There are more reads in "
                     "file 1 than in file 2.",