@marcelm 2021-12-16
Good improvement, thanks once again! I think the comment at the top of `__iter__` is a bit verbose, but let’s leave it as it is for now.